+++ src/applications/my-education-benefits/utils/validation.js
@@ -1,4 +1,6 @@
 import { isValidEmail } from 'platform/forms/validations';
+import moment from 'moment';
+import { formatReadableDate } from '../helpers';
 
 export const isValidPhone = value => {
   let stripped;
@@ -21,5 +23,22 @@ export const validatePhone = (errors, phone) => {
 export const validateEmail = (errors, email) => {
   if (email && !isValidEmail(email)) {
     errors.addError('Please enter a valid email address.');
+  }
+};
+
+export const validateEffectiveDate = (errors, dateString) => {
+  const effectiveDate = moment(dateString);
+  const minDate = moment().subtract(1, 'year');
+  const maxDate = moment().add(180, 'day');
+
+  if (
+    effectiveDate.isBefore(minDate, 'day') ||
+    effectiveDate.isAfter(maxDate, 'day')
+  ) {
+    errors.addError(
+      `Please enter a date between ${formatReadableDate(
+        minDate.format('YYYY-MM-DD'),
+      )} and ${formatReadableDate(maxDate.format('YYYY-MM-DD'))}`,
+    );
   }
 };
